face update
finally made some functions to draw the face and its weak spot (a third eye of sorts).  I also did rudimentary collision for the weak spot. I feel like there's a much easier and better way to do this but it works so im not complaining. the damage system works!