eye update
added working eyes that follow the player. the weak spot (which is also an eye) also has its own pupil that disappears when it's closed, and reappears when opened.

all i'm missing is the mouth, a pattern of attacks and movement (which will be semi random) and a different look for the player.