final update
i unfortunately had to omit the boss' movements and attacks as it caused a lot of bugs that made it unplayable and i dont have enough time to polish it.

So now all that works is player shooting, damaging the boss and there's a health bar that shows the boss' health.