third variation completed
completed the third variation, the fly clicker. i decided to leave the second variation last because it's the more complex one compared to the other 2. thankfully it seems like by changing the names of functions and variables is what fixes my issues with the menu template which makes sense. it's what i thought was the problem, but i made a little test earlier to see if it was but it still didn't work so i must've missed something.

this ones a mess HUD-wise, but i think it's fine