finished the 3rd variation and the jam in general
completed the frogger variation of the game. unfortunately it's a bit easy, but it works well in the end.

these games could use a bit of polish ESPECIALLY in the visual and code organising aspects, but I'm just glad they work fairly well with minimal to no bugs.

they're also a bit undercooked and simple in my opinion, but knowing me i'd waste way too much time trying to do too much so i'm trying to limit myself